Fixed up wrong data types in a few more API classes after DB rescaffold

+++ RestaurantAPI/RestaurantAPI.API/Models/Mapper.cs
@@ -34,8 +34,8 @@ namespace RestaurantAPI.Library
             Lat = other.Lat,
             Lon = other.Lon,
             Address = other.Address,
-            Rating = other.Rating,
-            PriceLevel = other.PriceLevel,
+            Rating = (decimal)other.Rating,
+            PriceLevel = (decimal)other.PriceLevel,
             Owner = other.Owner
         };
